requirements.txt
# Scraper dependencies
playwright>=1.40.0
pandas>=2.0.0
openpyxl>=3.1.0

# Uploader dependencies
mwclient>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
